src/modules/rentals/useCases/createRentals/CreateRentalsUseCase.ts
import { inject, injectable } from 'tsyringe'

import { Rental } from '@modules/rentals/infra/typeorm/entites/Rentals'

import { ICarsRepository } from '@modules/cars/repositories/ICarsRepository'
import { IRentalsRepository } from '@modules/rentals/repositories/IRentalsRepository'
import { IDateProvider } from '@shared/container/providers/DateProvider/IDateProvider'

import { AppError } from '@shared/errors/AppError'

interface IRequest {
  user_id: string
  car_id: string
  expected_return_date: Date
}

@injectable()
class CreateRentalsUseCase {
  constructor(
    @inject('RentalsRepository')
    private rentalsRepository: IRentalsRepository,
    @inject('DayjsDateProvider')
    private dateProvider: IDateProvider,
    @inject('CarsRepository')
    private carsRepository: ICarsRepository,
  ) {}

  async execute({
    user_id,
    car_id,
    expected_return_date,
  }: IRequest): Promise<Rental> {
    const minHourBetweenRentalAndExpectedReturn = 24
    const carUnavailable = await this.rentalsRepository.findOpenRentalByCar(
      car_id,
    )

    if (carUnavailable) {
      throw new AppError('Car is unavailable')
    }

    const rentalOpenToUser = await this.rentalsRepository.findOpenRentalByUser(
      user_id,
    )

    if (rentalOpenToUser) {
      throw new AppError('There is a rental in progress for user')
    }

    const compareDates = this.dateProvider.compareInHours(
      this.dateProvider.dateNow(),
      expected_return_date,
    )

    if (compareDates < minHourBetweenRentalAndExpectedReturn) {
      throw new AppError('Invalid return date and hour')
    }

    const rental = await this.rentalsRepository.create({
      user_id,
      car_id,
      expected_return_date,
    })

    await this.carsRepository.updateAvailable(car_id, false)

    return rental
  }
}

export { CreateRentalsUseCase }
